api-server-django/requirements.txt
asgiref==3.7.2
contourpy==1.1.1
cycler==0.12.1
deprecation==2.1.0
Django==3.2.13
django-3-jet==1.0.8
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
excel2img==1.4.0
fonttools==4.43.1
gunicorn==20.1.0
img2pdf==0.4.4
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.0
numpy==1.26.0
openpyxl==3.1.2
packaging==23.2
pikepdf==8.5.1
Pillow==10.0.1
plotly==5.18.0
psycopg2==2.9.9
PyJWT==2.4.0
pyparsing==3.1.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.8.0
